Fix bug with ctrl+z and ctrl+y not firing and prevent simultaneous undo+redo
