+++ src/serlo-editor/editor-ui/document-editor.tsx
@@ -35,7 +35,6 @@ export function DocumentEditor({
   hasToolbar,
 }: DocumentEditorProps) {
   const [hasHover, setHasHover] = useState(false)
-
   const editorStrings = useEditorStrings()
 
   const shouldShowSettings = showSettings()
